pyproject.toml
[tool.poetry]
name = "gendiff"
version = "0.1.0"
description = "Diff Generator"
authors = ["Denis Snegirev <snegirev.dv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pep8-naming = "^0.13.3"
flake8-return = "^1.2.0"
flake8-isort = "^6.1.1"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
coverage = "^7.4.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
